Allow MS.ANC.Razor.Serialization.Json to compile differently
This changes splits the ObjectReaders and ObjectWriters into partial classes that are compiled differently depending on the presence of two conditional compilation symbols:

- JSONSERIALIZATION_PROJECTSYSTEM - When enabled, compiles serialization for various project system types, such as RazorProjectInfo.
- JSONSERIALIZATION_ENABLETAGHELPERCACHE - When enabled, compiles ObjectReader methods that will return a TagHelperDescriptor from TagHelperCache if its checksum is already known.
