+++ src/Razor/src/Microsoft.VisualStudio.LanguageServices.Razor/Microsoft.VisualStudio.LanguageServices.Razor.csproj
@@ -9,6 +9,10 @@
     <ExcludeFromSourceOnlyBuild>true</ExcludeFromSourceOnlyBuild>
     <IsPackable Condition="'$(OS)' != 'Windows_NT'">false</IsPackable>
     <RootNamespace>Microsoft.VisualStudio.Razor</RootNamespace>
+
+    <!-- Live Share serializes ProjectSnapshotHandleProxy instance to JSON. So, we must include serialization
+         for project system types and enable TagHelperDescriptor caching. -->
+    <DefineConstants>$(DefineConstants);JSONSERIALIZATION_PROJECTSYSTEM;JSONSERIALIZATION_ENABLETAGHELPERCACHE</DefineConstants>
   </PropertyGroup>
 
   <ItemGroup>
