+++ src/Shared/Microsoft.AspNetCore.Razor.Serialization.Json/Microsoft.AspNetCore.Razor.Serialization.Json.projitems
@@ -16,9 +16,16 @@
     <Compile Include="$(MSBuildThisFileDirectory)JsonDataWriter.Policy.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)JsonReaderExtensions.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ObjectJsonConverter`1.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ObjectReaders_ProjectSystem.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ObjectReaders_TagHelpers.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ObjectReaders.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ObjectWriters_ProjectSystem.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ObjectWriters_TagHelpers.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ObjectWriters.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)SerializationFormat.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)StringCache.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)StringCache.Entry.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)StringCache.ExfiltratingEqualityComparer.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)WellKnownPropertyNames.cs" />
   </ItemGroup>
   <ItemGroup>
